backend/Date_Ocr/requirements.txt
cycler==0.10.0
decorator==4.4.1
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.2.0
django-filter==2.2.0
djangorestframework==3.10.3
gunicorn==20.0.4
imutils==0.5.3
kiwisolver==1.1.0
Markdown==3.1.1
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==6.2.1
psycopg2-binary==2.8.4
pytesseract==0.3.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
whitenoise==4.1.4
